SOLAR-Generator/requirements.txt

tqdm
matplotlib
ffmpeg-python

numpy >= 1.24
arcle >= 0.2.5
gymnasium >=0.28.0
